Bugfixes, collisions, and sprites
Fixed a bug where going outside of the road boundries caused the car to
stop. Fixed the car sprites, so it no longer looks like it's popping a
wheelie up hills.
Added crashing. when the car crashes into an object, it explodes for a
time relative to it's percentage of max speed. If you crash into another
car, your car is slowed significantly relative to your's and the other
car's speed.