New PlayerModule, game tweaks, etc.
wooo! player modules! check humanPlayer.js for a reference of how to use
it, because it uses inheritance from playerModule.js. With this, the
player data was moved around a bit, basically associating any car
physics to <player>.car, and meta about the player is stored on the base
level. Each player module has to at least overwrite the .accelerate(dt)
and .steer(dt) functions, and you might want to expand the constructor,
but it's not recommended that you overwrite anything else in
playerModule.

There are also a lost of new game tweaks thrown in. Other cars look a
little more realistic with slight model turning (although for now,
curves and z distance from player are not accounted for).